requirements.txt
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
gspread 
google-generativeai
python-dotenv
python-docx
pyPDF2
requests